- Fixed error in 32 bit LUT 3D - Seperated early termination into seperate functions

+++ test/libmorton_test.h
@@ -13,7 +13,7 @@
 #include "../libmorton/include/morton3D_64.h"
 
 // correct morton codes for 16 x 16 x 16, with z running fastest, then y, then x (4096 in total)
-static const int control_3D_Encode[4096] =
+static const uint_fast16_t control_3D_Encode[4096] =
 {
 	0, 4, 32, 36, 256, 260, 288, 292, 2048, 2052, 2080, 2084, 2304, 2308, 2336, 2340, 2,
 	6, 34, 38, 258, 262, 290, 294, 2050, 2054, 2082, 2086, 2306, 2310, 2338, 2342, 16,
